// 06032026/sce-front/src/app/model/reportes/filtro-comparacion-digitacion-asist-automa-.ts
export class FiltroComparacionDigitacionAsistAutoma {
  idProceso: number;
  idEleccion: number;
  idAmbito: number;
  idCentroComputo: number;
  proceso: string;
  eleccion: string;
  centroComputo: string;
  centroComputoCod: string;
  ubigeo: string;
  mesa: string;
  departamento: string;
  provincia: string;
  distrito: string;
}